PodcastEpisodeTemplate + clean up
